Annotation update
More migrations from nocedit

+++ annotations/nocturne.exe/pseudocode/src/FUN_00571ee0.c
@@ -2,11 +2,11 @@
 // Address: 00571ee0
 // Address Range: [[00571ee0, 00571eed]]
 // Convention: unknown
-// Signature: void FUN_00571ee0(undefined4 param_1)
+// Signature: void FUN_00571ee0(ulong param_1)
 
 #include "nocturne.h"
 
-void FUN_00571ee0(uint param_1)
+void FUN_00571ee0(ulong param_1)
 
 {
   malloc(param_1);
