+++ annotations/nocturne.exe/pseudocode/src/sound/sndmain.cpp/getMaxSwLatency_FUN_00528970.cpp
@@ -2,11 +2,11 @@
 // Address: 00528970
 // Address Range: [[00528970, 00528975]]
 // Convention: __cdecl
-// Signature: undefined4 __cdecl sound_sndmain_cpp_getMaxSwLatency_FUN_00528970(void)
+// Signature: float __cdecl sound_sndmain_cpp_getMaxSwLatency_FUN_00528970(void)
 
 #include "nocturne.h"
 
-uint __cdecl sound_sndmain_cpp_getMaxSwLatency_FUN_00528970(void)
+float __cdecl sound_sndmain_cpp_getMaxSwLatency_FUN_00528970(void)
 
 {
   return DAT_005bea98;
